fix: resolve all clippy warnings
- Remove redundant ToPrimitive imports in traits.rs
- Use RangeInclusive::contains instead of manual bounds checks
- Collapse nested if statements
- Use *= operator for assign operations
- Remove unnecessary type casts
- Use Option::map instead of if-let-some/else-none
- Use if-let instead of single-arm match

+++ src/js_semantics.rs
@@ -87,7 +87,7 @@ impl NumericValue {
                 // Convert rational to integer (truncate)
                 r.to_integer()
                     .to_i32()
-                    .unwrap_or_else(|| (r.to_integer() as i64) as i32)
+                    .unwrap_or_else(|| r.to_integer() as i32)
             }
             NumericValue::Decimal(d) => {
                 // Try direct conversion first (fast path)
